Foto dos produtos

+++ sistema/upload.php
@@ -13,11 +13,11 @@
          
     if(isset($_FILES)){
 		
-		//print_r($_FILES);
+		//print_r($_FILES['imagem']);die;
 		
-        $nome_imagem    = $_FILES['file']['name'];
-        $tamanho_imagem = $_FILES['file']['size'];
-         
+        $nome_imagem    = $_FILES['imagem']['name'];
+        $tamanho_imagem = $_FILES['imagem']['size'];
+		
         // pega a extensão do arquivo
         $ext = strtolower(strrchr($nome_imagem,"."));
          
@@ -30,7 +30,7 @@
             if($tamanho < 1024){ //se imagem for até 1MB envia
                 $nome_atual = md5(uniqid(time())).$ext; //nome que dará a imagem
 
-                $tmp = $_FILES['file']['tmp_name']; //caminho temporário da imagem
+                $tmp = $_FILES['imagem']['tmp_name']; //caminho temporário da imagem
                 
                 $destino = $pasta.$nome_atual;
 
@@ -41,15 +41,19 @@
                 if(move_uploaded_file($tmp, $destino)){
                     
                     //verifica se já foi adicionado alguma foto, se sim, exclui ela fisicamente, já que nao será mais usada, eliminando lixo.
-                    if ($_SESSION['fotoAtual'] != ''){
-						unlink($pasta.$_SESSION['fotoAtual']);
+                    if (isset($_SESSION['fotoAtual']) and $_SESSION['fotoAtual'] != ''){
+						if (file_exists($pasta.$_SESSION['fotoAtual'])){
+							unlink($pasta.$_SESSION['fotoAtual']);
+						}
 					}                    
                     
                     echo "<a href=\"".$pasta.$nome_atual."\" class=\"fancybox\"><img src='".$pasta.$nome_atual."' class='ml-3' style='max-width: 260px; max-height:250px; border:2px solid #ccc;'></a>"; //imprime a foto na tela
                     
-                    $_SESSION['fotoAtual'] = $nome_atual; //guarda o nome da foto pra gravar no banco
+                    $_SESSION['fotoAtual'] = $nome_atual; //guarda o nome da foto pra gravar no banco					
                 }else{
                     echo "Falha ao enviar";
+					
+					$_SESSION['fotoAtual'] = '';
                 }
             }else{
                 echo "A imagem deve ser de no máximo 1MB";
